code/requirements.txt
python-dotenv
torch
wandb
